requirements-dev.txt
# Development dependencies

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-docker==3.1.0
pytest-env==1.1.5
pytest-cov==5.0.0

# Code quality
black==24.4.0
flake8==7.0.0
mypy==1.11.0
isort==5.13.0

# Documentation
sphinx==7.3.0
sphinx-rtd-theme==2.0.0

# Debugging
ipdb==0.13.13
ipython==8.26.0